+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Added
 * SSL is now supported if `endpoint_url` uses the `https` scheme (uses ruby-2.1 syntax internally)
+* New config: set `ssl_no_verify` to `true` to bypass SSL certificate verification.
+  Use at your own risk.
 ### Changed
 * Fixed tests:
   * Removed some warnings